+++ extensions/battery-health/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Battery Health Check Changelog
 
+## [Added Battery Health Percentage] - {PR_MERGE_DATE}
+
+ - Extension now shows the health of your battery relative to when it was new
+
 ## [Add Battery Health Check] - 2022-07-05
 
 Initial version code